chore(deps): update all deps
